Funcional version witouth id of gas_station validation

+++ backend/habragasolina/src/main/java/bo/edu/usfa/gasolina/habragasolina/Repository/GasStationRepository.java
@@ -7,6 +7,6 @@ import bo.edu.usfa.gasolina.habragasolina.Entities.GasStation;
 
 
 @Repository
-public interface GasStationRepository extends JpaRepository <GasStation, Long> {
+public interface GasStationRepository extends JpaRepository <GasStation, Integer> {
 
 }
